src/01-dinov3/requirements.txt
torch>=2.9
torchvision>=0.24
transformers>=5.3
tensorboard>=2.20
pillow>=12.0
scikit-learn>=1.8
numpy>=2.4
pandas>=2.3
onnx>=1.16
onnxruntime-gpu>=1.18
skl2onnx>=1.17
optimum>=1.21
